\name{logLik.mlcm}
\alias{logLik.mlcm}

\title{Extract Log-Likelihood from mlcm Object }
\description{
 This is a method function for extracting the log likelihood from objects of class \sQuote{mlcm}.
}
\usage{
\method{logLik}{mlcm}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object, \dots}{ object of class \code{mlcm} typically generated by \code{mlcm}. }
}
\details{
 See the help page for the generic function \code{\link{logLik} for details.}
}
\value{
  Returns an object of class \code{logLik} with an attribute, \code{df}, giving the degrees of freedom.
    }

\author{ Kenneth Knoblauch }

\seealso{  \code{\link{logLik}}}
\examples{
 logLik(mlcm(BumpyGlossy))
}
\keyword{ models }
\keyword{ regression }
